Add files via upload
These compressed folders contain only the sketch files that have been modified from the general code folder provided. In order for these to run, the sketch files in the original code folder with the p5 sound libraries and assets will need to be replaced with the sketch files I have shared.